requirements.txt
appnope==0.1.0
autopep8==1.3.3
base58==0.2.5
beautifulsoup4==4.6.0
bleach==2.0.0
browsercookie==0.7.2
bs4==0.0.1
certifi==2018.1.18
cfscrape==1.8.0
chardet==3.0.4
coinmarketcap==4.2.1
coverage==4.4.2
cycler==0.10.0
decorator==4.1.2
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.11.2
django-allauth==0.35.0
-e git://github.com/amarandon/django-audiotracks.git@e2480ebe555b07cc3c3c60b075a7caed462ed96d#egg=django_audiotracks
django-cors-headers==2.2.0
django-crispy-forms==1.6.1
django-filter==1.1.0
django-rest-auth==0.9.3
django-rest-multiple-models==2.0.1
django-webpack-loader==0.6.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
docker==2.7.0
docker-pycreds==0.2.1
drf-writable-nested==0.4.1
entrypoints==0.2.3
flake8==3.5.0
future==0.14.3
gunicorn==19.7.1
html2text==2016.9.19
html5lib==0.999999999
idna==2.6
ijson==2.3
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==7.0.0
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
keyring==10.4.0
line-profiler==2.0
lxml==3.8.0
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.0.2
mccabe==0.6.1
MechanicalSoup==0.7.0
mistune==0.7.4
mutagen==1.39
nbconvert==5.3.0
nbformat==4.4.0
notebook==5.0.0
numpy==1.14.0
oauthlib==2.0.6
olefile==0.44
pandas==0.22.0
pandocfilters==1.4.2
path.py==11.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.1
plotly==2.4.1
pluggy==0.6.0
prompt-toolkit==1.0.15
psycopg2==2.7.1
ptyprocess==0.5.2
py==1.5.2
py-postgresql==1.2.1
pyasn1==0.3.7
pyblake2==1.1.0
pycodestyle==2.3.1
pycrypto==2.6.1
pydub==0.20.0
PyExecJS==1.4.0
pyflakes==1.6.0
Pygments==2.2.0
PyJWT==1.6.0
pymongo==3.4.0
pyparsing==2.2.0
PySocks==1.6.8
pyTelegramBotAPI==3.1.0
python-axolotl-curve25519==0.1
python-dateutil==2.6.1
python-linkedin==4.2
python3-openid==3.1.0
pytz==2017.2
PyWaves==0.8.1
pyzmq==16.0.2
qtconsole==4.3.1
requests==2.18.4
requests-cache==0.4.13
requests-oauthlib==0.8.0
rsa==3.4.2
scipy==0.19.1
selenium==3.7.0
simplegeneric==0.8.1
six==1.10.0
steampy==0.46
stem==1.6.0
terminado==0.6
testpath==0.3.1
tornado==4.5.2
tox==2.9.1
traitlets==4.3.2
ulib==0.26
urllib3==1.22
virtualenv==15.1.0
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.46.0
widgetsnbextension==3.0.2
xlwt==1.3.0
